motion_field_projection: fold operations
Computing the projection of one coordinate is sufficient to determine
whether the block position will be invalid.

In addition, a value of 0 is not unusual for a MV coordinate, which
greatly simplifies the computations.

On 2 samples, cycle count of the codeblock goes 470k->420k and
620k->560k, so about a 10% speedup of the function. These sequences
may have had an usual amount of static scenes to explain this
phenomenon, though.