add_tpl_ref_mv: use cache for last motion result
This is not so unfrequent that this last motion is identical, therefore
allowing to reuse the computed motion and stack index, and skipping
some checks (as the motion is already in the stack now).

However, more than one item in the cache, whatever its origin (first to
last, first column, (0,0), ...), does not help, or even causes an
important slowdown).

Timing goes eg from 623 to 513. Overall 0.2% speedup.