Early exit for dav1d_init_ref_mv_tile_row
A call to motion_field_projection can be overwritten by the next one.
If this next one completely fills the buffer, then any preceeding one
becomes useless. However, it may happens that some holes in one field
are filled by a previous call.

Furthermore, within a call, a MI cell can overwrite the result
for a previous one. Therefore, by:
1) Completely reversing the processing order
2) Check if already set
it can determined when the motion field for the tile/frame has been
completely processed.

Therefore, to do early exit:
1) Reverse the processing order of MI cells and of ref frames
2) Count the number of cells changed
3) Exit if all done
4) Otherwise, go to the next ref frames

If done per ref frame, skips between 8% and 25% of ref frames.
If done per MI cell, skips from 20% to 40% of them.
However, a complete iteration for a ref frame has been measured
to be 15-25%% slower than the original one, so the overall impact
can be negative.

From 0.2% to 0.7% overall speedup.