recon_tmpl: use precomputed info for coeff parsing
The precomputed info concerns the raster scan address, the offset into
the level context array (instead of coordinates) and the offsets used
in computing the contexts for below 4 and below 16 coeffs.

While this is trading cache for instructions, even on heavy loads, the
trade-off is positive. Padding the struct to 8 bytes (or making some
use of the additional bytes) did not seem worthwhile.